requirements.txt
anthropic>=0.40.0
openai>=1.50.0
replicate>=0.30.0
celery[redis]>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0
pgvector>=0.2.4
Pillow>=10.0.0
boto3>=1.34.0
python-slugify>=8.0.4
requests>=2.31.0
python-dotenv>=1.0.0
sentry-sdk[celery]>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
tenacity>=8.2.0
httpx>=0.27.0
